api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["sepehr masihpour"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.116.1"
uvicorn = "^0.35.0"
web3 = "^7.12.1"
python-dotenv = "^1.1.1"
pydantic = "^2.11.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
